+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning].
 - RandomSkunk.Results:
     - Add overloads for enumerable extensions.
     - Add constructor to `Error` - makes initializing an error with extension properties easier.
+    - Add setting to catch exceptions thrown by callbacks (defaults to true).
 
 - RandomSkunk.Results.AspNetCore:
     - Add `GetActionResult` extension method for `Error`.